requirements.txt
APScheduler==3.6.3
cachetools==4.2.2
certifi==2022.12.7
et-xmlfile==1.1.0
numpy==1.23.5
openpyxl==3.0.10
pandas==1.5.2
python-dateutil==2.8.2
python-dotenv==0.21.0
python-telegram-bot==13.15
pytz==2022.6
pytz-deprecation-shim==0.1.0.post0
six==1.16.0
tornado==6.1
tzdata==2022.7
tzlocal==4.2
